+++ src/models/purchase/purchase.model.js
@@ -6,8 +6,8 @@ module.exports = model;
 
 function model(sequelize) {
     const attributes = {
-        id: { type: DataTypes.BIGINT, autoIncrement: true, primaryKey: true },
-        fk_provider_id: { type: DataTypes.BIGINT, allowNull: false },
+        id: { type: DataTypes.INTEGER, autoIncrement: true, primaryKey: true },
+        fk_provider_id: { type: DataTypes.INTEGER, allowNull: false },
         expected_arrival_date: { type: DataTypes.DATE, allowNull: false },
         total_amount: { type: DataTypes.DECIMAL(18, 0), allowNull: false },
         status: { type: DataTypes.TINYINT, allowNull: false },
