+++ src/models/user/profile.model.js
@@ -6,7 +6,7 @@ module.exports = model;
 
 function model(sequelize) {
     const attributes = {
-        id: { type: DataTypes.BIGINT, autoIncrement: true, primaryKey: true },
+        id: { type: DataTypes.INTEGER, autoIncrement: true, primaryKey: true },
         fk_user_code: { type: DataTypes.STRING(20), allowNull: false },
         first_name: { type: DataTypes.STRING(255), allowNull: false },
         last_name: { type: DataTypes.STRING(255), allowNull: false },
